👷 Increase CI granularity
This will also fix an issue with codecov where pushing to main has
resulted in codecov comparing against a *very* old commit, even if a
pull request commit had higher stats shown.

We will eventually be able to do more steps/requirements in the PR
workflow after this is merged as well.

Signed-off-by: Isabella Muerte <63051+slurps-mad-rips@users.noreply.github.com>